+++ Modules/skinning.lua
@@ -112,6 +112,50 @@ local function ApplyCooldownFont(cooldownFrame, options)
 			cooldownFontString:SetFont(unpack(originalCooldownFont))
 		end
 	end
+
+	local parent = cooldownFrame:GetParent()
+	if parent.SCMConfig then
+		cooldownFrame:SetHideCountdownNumbers(parent.SCMConfig.hideCountdownNumbers)
+	end
+end
+
+local function ApplyCooldownSwipe(cooldownFrame, options)
+	local parent = cooldownFrame:GetParent()
+	local forceActiveSwipe = parent.SCMConfig and parent.SCMConfig.forceActiveSwipe
+
+	if parent.auraInstanceID or parent.SCMFakeAuraInstanceID or parent.SCMBuffOptions then
+		if options.disableRegularIconActiveSwipe and not forceActiveSwipe then
+			if options.recolorNormalSwipe then
+				cooldownFrame:SetSwipeColor(unpack(options.normalSwipeColor))
+			else
+				cooldownFrame:SetSwipeColor(0, 0, 0, 0.7)
+			end
+
+			if parent.SCMBuffOptions then
+				cooldownFrame:SetReverse(options.reverseActiveSwipe)
+			end
+		else
+			if options.recolorActiveSwipe then
+				cooldownFrame:SetSwipeColor(unpack(options.activeSwipeColor))
+			end
+
+			cooldownFrame:SetReverse(options.reverseActiveSwipe)
+		end
+	elseif options.recolorNormalSwipe then
+		cooldownFrame:SetSwipeColor(unpack(options.normalSwipeColor))
+		cooldownFrame:SetReverse(false)
+	else
+		cooldownFrame:SetSwipeColor(0, 0, 0, 0.7)
+	end
+end
+
+local function OnSetCooldown(self)
+	local options = SCM.db.profile.options
+
+	SCM.Cooldowns.ApplyNumericRuleFormatter(self)
+
+	ApplyCooldownSwipe(self, options)
+	ApplyCooldownFont(self, options)
 end
 
 local function ApplyCooldownStyle(child, options)
@@ -126,46 +170,16 @@ local function ApplyCooldownStyle(child, options)
 			child.CooldownFlash:SetAlpha(0)
 		end
 
-		cooldownFrame:ClearAllPoints()
-		cooldownFrame:SetPoint("TOPLEFT", child, "TOPLEFT", 0, -0)
-		cooldownFrame:SetPoint("BOTTOMRIGHT", child, "BOTTOMRIGHT", -0, 0)
+		child.Cooldown:ClearAllPoints()
+		child.Cooldown:SetAllPoints()
+		-- if child.SCMCustom then
+		-- 	cooldownFrame:SetPoint("TOPLEFT", child, "TOPLEFT", 0, -SCM:PixelPerfect())
+		-- end
+		cooldownFrame:SetPoint("BOTTOMRIGHT", child, "BOTTOMRIGHT", -SCM:PixelPerfect(), SCM:PixelPerfect())
 		cooldownFrame:SetSwipeTexture("Interface\\Buttons\\WHITE8x8")
 
-		hooksecurefunc(cooldownFrame, "SetCooldown", function(self)
-			local options = SCM.db.profile.options
-			local parent = self:GetParent()
-			local forceActiveSwipe = parent.SCMConfig and parent.SCMConfig.forceActiveSwipe
-
-			SCM.Cooldowns.ApplyNumericRuleFormatter(self)
-
-			if parent.auraInstanceID or parent.SCMFakeAuraInstanceID or parent.SCMBuffOptions then
-				if options.disableRegularIconActiveSwipe and not forceActiveSwipe then
-					if options.recolorNormalSwipe then
-						self:SetSwipeColor(unpack(options.normalSwipeColor))
-					else
-						self:SetSwipeColor(0, 0, 0, 0.7)
-					end
-
-					if parent.SCMBuffOptions then
-						self:SetReverse(options.reverseActiveSwipe)
-					end
-				else
-					if options.recolorActiveSwipe then
-						self:SetSwipeColor(unpack(options.activeSwipeColor))
-					end
-
-					self:SetReverse(options.reverseActiveSwipe)
-				end
-			elseif options.recolorNormalSwipe then
-				self:SetSwipeColor(unpack(options.normalSwipeColor))
-				self:SetReverse(false)
-			else
-				self:SetSwipeColor(0, 0, 0, 0.7)
-			end
-			ApplyCooldownFont(self, options)
-		end)
-
-		ApplyCooldownFont(cooldownFrame, options)
+		hooksecurefunc(cooldownFrame, "SetCooldown", OnSetCooldown)
+		OnSetCooldown(cooldownFrame)
 	end
 end