+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-### Version 1.5.29 [Alpha]
+### Version 1.5.29 [Release]
 
 Additions:
+- Add option to hide the timer text
+- Increase the maximum value for the cast bar width
+- Add an option to load an icon when a spell is NOT known
+- 
 
 Fixes:
 - Fix pets again (hopefully)
 - Fix pandemic glow again (hopefully)
 - Fix "Desaturate While Inactive" (hopefully)
+- Fix some import strings
+- Fix panda and dracthyr load condition
+- Hopefully fix the pandemic glow
+- Fix Healthstone charges
+- Fix Mindbender
+- Fix the display style setting of the cooldown text formatting
+- Hopefully fix the pixel perfect issues of the cooldown swipe
+- Fix "Hide When Ready" for spells with charges